// frontend/src/api/api.ts
export interface LoginDto {
    /**
     * The email address of the user
     * @type {string}
     * @memberof LoginDto
     */
    'email': string;
    /**
     * The password of the user
     * @type {string}
     * @memberof LoginDto
     */
    'password': string;
}
/**
 * 
 * @export
 * @interface LoginResponseDto
 */
export interface LoginResponseDto {
    /**
     * 
     * @type {User}
     * @memberof LoginResponseDto
     */
    'user': User;
    /**
     * 
     * @type {string}
     * @memberof LoginResponseDto
     */
    'accessToken': string;
}
// ...
export interface User {
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'phone': string;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    'age': number;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'password': string;
}